requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
motor>=3.3.0
pymongo>=4.6.0
boto3>=1.34.0
python-decouple>=3.8
celery>=5.3.0
redis>=5.0.0
ffmpeg-python>=0.2.0
aiofiles>=23.2.0
Pillow>=10.0.0
pydantic>=2.5.0
httpx>=0.25.0
